Refactor addon to run as dbus service
Anaconda moves to executing actual tasks in separate processes and
communicate with them via dbus.

Refactor the addon to the new model, based on documentation in
https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/9/html-single/customizing_anaconda/index#developing-installer-add-ons_customizing-anaconda
and
https://github.com/rhinstaller/hello-world-anaconda-addon

Split the addon into few modules, as required by the new model:
 - kickstart - for serializing/deserializing kickstart content
 - main service
 - dbus interface wrapper for the main service

This extra abstraction layer requires defining all properties 3 times.
Furthermore, D-Bus interface requires all names to be CamelCase. Handle
this duplication automatically as much as possible, but unfortunately
some is unavoidable.

The D-Bus name must be within org.fedoraproject.Anaconda.Addons
namespace, otherwise it is ignored by initial-setup.

Besides code move, this commit tries to not alter behavior of the tasks
themselves.

QubesOS/qubes-issues#6982
