+++ dbus/org_qubes_os_initial_setup.conf
@@ -1,0 +1,13 @@
+<!DOCTYPE busconfig PUBLIC
+"-//freedesktop//DTD D-BUS Bus Configuration 1.0//EN"
+"http://www.freedesktop.org/standards/dbus/1.0/busconfig.dtd">
+<busconfig>
+       <policy user="root">
+               <allow own="org.fedoraproject.Anaconda.Addons.QubesInitialSetup"/>
+               <allow send_destination="org.fedoraproject.Anaconda.Addons.QubesInitialSetup"/>
+       </policy>
+       <policy context="default">
+               <deny own="org.fedoraproject.Anaconda.Addons.QubesInitialSetup"/>
+               <allow send_destination="org.fedoraproject.Anaconda.Addons.QubesInitialSetup"/>
+       </policy>
+</busconfig>
